+++ src/app/tests/sagaTester.test.js
@@ -909,6 +909,56 @@ Error: ERROR`);
       const expectedError = 'no matching set of parameters were found!';
       expect(() => new SagaTester(saga, config).run(action)).toThrow(expectedError);
     });
+    it('should correctly use subsequent entries in the expectedCalls list if multiple candidates exist', () => {
+      const method1 = () => {};
+      const method2 = () => {};
+
+      function* saga() {
+        const results = [];
+        results.push(yield call(method1));
+        results.push(yield call(method1));
+        results.push(yield call(method1));
+
+        results.push(yield call(method2, 'arg'));
+        results.push(yield call(method2, 'arg'));
+        results.push(yield call(method2, 'arg'));
+
+        return results;
+      }
+
+      expect(new SagaTester(saga, {
+        expectedCalls: [
+          { name: 'method2', times: 2, params: ['arg'], output: 'method2-1' },
+          { name: 'method1', output: 'method1-1' },
+          { name: 'method1', times: 1, params: [], output: 'method1-2' },
+          { name: 'method1', output: 'method1-3' },
+          { name: 'method2', params: ['arg'], output: 'method2-2' },
+        ],
+      }).run()).toEqual([
+        'method1-1',
+        'method1-2',
+        'method1-3',
+        'method2-1',
+        'method2-1',
+        'method2-2',
+      ]);
+    });
+    it('should fail if there is a times: 0 element in expectedCalls', () => {
+      const method1 = () => {};
+
+      function* saga() {
+        yield call(method1);
+      }
+
+      expect(() => new SagaTester(saga, {
+        expectedCalls: [
+          { name: 'method1', times: 0, output: 'method1-1' },
+          { name: 'method1', output: 'method1-1' },
+          { name: 'method1', times: 1, params: [], output: 'method1-2' },
+          { name: 'method1', output: 'method1-3' },
+        ],
+      }).run()).toThrow('Expected to receive 0 calls to method1. Received 1');
+    });
   });
 
   describe('handling inner generator methods', () => {